requirement.txt
gradio
loguru
ultralytics
pillow
